Bill Interface completed
